server/requirements.txt
fastapi
uvicorn
whisper
torch
google-generativeai
python-dotenv